requirements.txt
openai==1.66.3
pinecone==6.0.2
python-dotenv==1.0.1
streamlit==1.39.0
